+++ UI/Library/Controls/SceneEditing/IGizmo.cs
@@ -1,7 +1,7 @@
 ﻿namespace Macabre2D.UI.Library.Controls.SceneEditing {
 
-    using Macabre2D.UI.Library.Models.FrameworkWrappers;
     using Macabre2D.Framework;
+    using Macabre2D.UI.Library.Models.FrameworkWrappers;
     using Microsoft.Xna.Framework;
     using Microsoft.Xna.Framework.Input;
 
@@ -11,7 +11,7 @@
 
         void Draw(FrameTime frameTime, BoundingArea viewBoundingArea, BaseComponent selectedComponent);
 
-        void Initialize(EditorGame game);
+        void Initialize(SceneEditor game);
 
         bool Update(FrameTime frameTime, MouseState mouseState, KeyboardState keyboardState, Vector2 mousePosition, ComponentWrapper selectedComponent);
     }