+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Merged editor and framework back into a combined repository.
 * Project autosaves now go into a special hidden folder.
 * Time handled through new TimeFrame object, which will allow adjustments to game speed.
+* Using MonoGame.WpfCore template instead of MonoGame.Framework.WpfInterop to display and edit scenes in the editor.
 
 ### Fixes